@@ -1,26 +1,22 @@
-﻿namespace Vapolia.FluentLayouts.Extensions
+﻿namespace Vapolia.FluentLayouts.Extensions;
+
+internal static class ArrayExtensions
 {
-	internal static class ArrayExtensions
+	public static bool TryGetElement<T>(this T[] array, int index, out T element)
 	{
-		public static bool TryGetElement<T>(this T[] array, int index, out T element)
+		if (array == null)
 		{
-			if (array == null)
-			{
-				element = default(T);
-
-				return false;
-			}
-
-			if (index < array.Length)
-			{
-				element = array[index];
-
-				return true;
-			}
-
-			element = default(T);
-
+			element = default;
 			return false;
 		}
+
+		if (index < array.Length)
+		{
+			element = array[index];
+			return true;
+		}
+
+		element = default;
+		return false;
 	}
 }
