@@ -1,5 +1,4 @@
 using System.Runtime.InteropServices;
-using UIKit;
 
 namespace Vapolia.FluentLayouts;
 
@@ -15,11 +14,11 @@ public class UIViewAndLayoutAttribute
     public NSLayoutAttribute Attribute { get; }
 
     public FluentLayout EqualTo(NFloat constant = default) =>
-        new FluentLayout(View, Attribute, NSLayoutRelation.Equal, constant);
+        new (View, Attribute, NSLayoutRelation.Equal, constant);
 
     public FluentLayout GreaterThanOrEqualTo(NFloat constant = default) =>
-        new FluentLayout(View, Attribute, NSLayoutRelation.GreaterThanOrEqual, constant);
+        new (View, Attribute, NSLayoutRelation.GreaterThanOrEqual, constant);
 
     public FluentLayout LessThanOrEqualTo(NFloat constant = default) =>
-        new FluentLayout(View, Attribute, NSLayoutRelation.LessThanOrEqual, constant);
+        new (View, Attribute, NSLayoutRelation.LessThanOrEqual, constant);
 }
